# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Course Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/students:
    get:
      operationId: listAdminStudents
      tags: [admin]
      summary: List all students with admin details
      responses:
        "200":
          description: Students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminStudent"

    post:
      operationId: createAdminStudent
      tags: [admin]
      summary: Create a new user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminStudentCreate"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
        "409":
          description: Email or phone already exists
  /admin/students/{id}:
    get:
      operationId: getAdminStudent
      tags: [admin]
      summary: Get a student profile (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
        "404":
          description: Not found
    patch:
      operationId: updateAdminStudent
      tags: [admin]
      summary: Update a student profile (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminStudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudent"
    delete:
      operationId: deleteAdminStudent
      tags: [admin]
      summary: Delete a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /auth/register:
    post:
      operationId: authRegister
      tags: [users]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLogin
      tags: [users]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Device limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceLimitError"
  /auth/remove-device:
    post:
      operationId: authRemoveDevice
      tags: [users]
      summary: Remove a device session using credentials (no auth cookie required)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRemoveDeviceInput"
      responses:
        "204":
          description: Device removed
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Device not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
    Announcement:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
    UserDevice:
      type: object
      required: [id, userId, platform, loginAt, lastActiveAt, current]
      properties:
        id:
          type: integer
        userId:
          type: integer
        platform:
          type: string
        userAgent:
          type: ["string", "null"]
        loginAt:
          type: string
          format: date-time
        lastActiveAt:
          type: string
          format: date-time
        current:
          type: boolean
          description: Whether this is the device making the current request
...
    User:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        studentId:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, instructor, admin, mentor, editor]
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        examTitle:
          type: ["string", "null"]
        institute:
          type: ["string", "null"]
        passingYear:
          type: ["string", "null"]
        cvLink:
          type: ["string", "null"]
        githubLink:
          type: ["string", "null"]
        portfolioLink:
          type: ["string", "null"]
        linkedinLink:
          type: ["string", "null"]
        professionalImageLink:
          type: ["string", "null"]
...
    AdminStudent:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        studentId:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, mentor, editor, admin, instructor]
        enrolledCount:
          type: integer
        createdAt:
          type: string
        status:
          type: ["string", "null"]
        addressLine1:
          type: ["string", "null"]
        addressLine2:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        zipCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        examTitle:
          type: ["string", "null"]
        institute:
          type: ["string", "null"]
        passingYear:
          type: ["string", "null"]
        cvLink:
          type: ["string", "null"]
        githubLink:
          type: ["string", "null"]
        portfolioLink:
          type: ["string", "null"]
        linkedinLink:
          type: ["string", "null"]
        professionalImageLink:
          type: ["string", "null"]
...
    AdminStudentCreate:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, mentor, editor, admin]
    AdminStudentUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        address:
          type: string
        phone:
          type: string
        bio:
          type: string
        avatar:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, mentor, editor, admin]
        status:
          type: string
          enum: [active, suspended]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AuthRegisterInput:
      type: object
      required: [name, email, phone, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
          minLength: 6
        password:
          type: string
          minLength: 6
    AuthLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    DeviceLimitError:
      type: object
      required: [error, devices]
      properties:
        error:
          type: string
        devices:
          type: array
          items:
            $ref: "#/components/schemas/UserDevice"
    AuthRemoveDeviceInput:
      type: object
      required: [email, password, deviceId]
      properties:
        email:
          type: string
        password:
          type: string
        deviceId:
          type: integer